TLRC - 826 - Technical Story - POC for Registrations file - Fixed issue with IFormFile

+++ Sfa.Poc.ResultsAndCertification.CsvHelper/Sfa.Poc.ResultsAndCertification.CsvHelper.Domain/Models/TqRegistration.cs
@@ -11,7 +11,7 @@ namespace Sfa.Poc.ResultsAndCertification.CsvHelper.Domain.Models
             TqSpecialismRegistrations = new HashSet<TqSpecialismRegistration>();
         }
 
-        public int UniqueLearnerNumber { get; set; }
+        public long UniqueLearnerNumber { get; set; }
         public string Firstname { get; set; }
         public string Lastname { get; set; }
         public DateTime?  DateofBirth { get; set; }
